tset_repo/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6

python-jose==3.3.0
python-jose[cryptography]
python-decouple=3.6

pymongo==3.12.1

passlib==1.7.4
passlib[bcrypt]